Change prefix for blueprint reserved keys
Previously _classname and _module where reserved keys in the yaml
file used for object instantiation if build is called on the blueprint.
However, _module and _classname are not totally improbable variable
names. The choice of $classname and $module is probably less likely
to have collisions since $classname is not a valid python variable,
but can be stored in a class with setattr with no issues.
Moreover, $ is not used in yaml, so parsing the file should still
work without issues.