Drop support for py27 - py36
In these versions of python, the order of keys in a dictionary cannot be
relied upon. This creates the following issue for Blueprints: if we want
to guarantee that the same dict will be serialized to the same yaml - we
need to first sort the keys when we create the Blueprint and then use
the default pyyaml dump that sorts the keys before writing to the file.

From python3.6 and onwards, however, dicts are guaranteed to maintain
their keys in insertion order. This, coupled with the fact that PyYaml
now has an option to dump without sorting the keys, at last gives us the
nice feature of our Blueprints retaining the semantic structure we had
in mind when we wrote them in a particular order. Dropping support
for earlier versions of python is annoying, but I really want this
feature.